requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=1.0.0
Pillow>=9.4.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
xgboost>=2.0.0
openai>=1.0.0
tqdm>=4.60.0
pyyaml>=6.0
